test(ui): split the regex search tests into their own file
model_test.go had grown larger than any source file in the package. Move the
search tests alongside the source split, taking their shared fixture with them.

Tests move unchanged.
